refactors

+++ db/user_service.py
@@ -41,5 +41,5 @@ def set_salary_by_id(user_id: int, user: User) -> None:
 
 def remove_user_by_id(user_id: int) -> User:
     conn = sqlite3.connect('my_app.db')
-    conn.execute("DELETE FROM user WHERE id = ?")
+    conn.execute("DELETE FROM user WHERE id = ?", (user_id))
